requirements.txt
fastapi
uvicorn
selenium
undetected-chromedriver
gspread
google-auth
python-dotenv
pytz
requests
setuptools
